requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
plotly>=5.18.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
supabase>=2.0.0
